edit Dashboard
